python-scripts/requirements.txt
boto3==1.36.26
botocore==1.36.26
fastavro==1.10.0
jmespath==1.0.1
psycopg2==2.9.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
s3transfer==0.11.2
six==1.17.0
urllib3==2.3.0
